Adicionei as funções dos botões
