+++ src/components/header/Stl-Header.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+
+export const HeaderContainer = styled.header`
+display: flex;
+justify-content: space-between;
+background-color: orange;
+height: 50px;
+align-items: center;
+padding: 0 1rem;
+nav button:nth-child(1){
+    margin: 0 3px 0 0;
+
+}
+`
