+++ src/components/productsContainer/ProductsContainer.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import ProductCard from '../productCard/ProductCard'
-import { StyledContainerProduct } from './styled.productsContainer'
+import { StyledContainerProduct } from './styled-productsContainer'
 import Filters from './filters/Filters'
 
 export default class ProductsContainer extends Component {
